+++ pipeline/join.go
@@ -84,6 +84,42 @@ type JoinNode struct {
 	//   - none - (default) skip rows where a point is missing, inner join.
 	//   - null - fill missing points with null, full outer join.
 	//   - Any numerical value - fill fields with given value, full outer join.
+	//
+	// When using a numerical or null fill, the fields names are determined by copying
+	// the field names from another point.
+	// This doesn't work well when different sources have different field names.
+	// Use the DefaultNode and DeleteNode to finalize the fill operation if necessary.
+	//
+	// Example:
+	//    var maintlock = stream
+	//        |from()
+	//            .measurement('maintlock')
+	//            .groupBy('service')
+	//    var requests = stream
+	//        |from()
+	//            .measurement('requests')
+	//            .groupBy('service')
+	//    // Join the maintlock and requests streams
+	//    // The intent it to drop any points in maintenance mode.
+	//    maintlock
+	//        |join(requests)
+	//            // Provide prefix names for the fields of the data points.
+	//            .as('maintlock', 'requests')
+	//            // points that are within 1 second are considered the same time.
+	//            .tolerance(1s)
+	//            // fill missing fields with null, implies outer join.
+	//            // a better default per field will be set later.
+	//            .fill('null')
+	//            // name the resulting stream.
+	//            .streamName('requests')
+	//        |default()
+	//            // default maintenance mode to false, overwriting the null value if present.
+	//            .field('maintlock.mode', false)
+	//            // default the requests to 0, again overwriting the null value if present.
+	//            .field('requests.value', 0.0)
+	//        // drop any points that are in maintenance mode.
+	//        |where(lambda: "maintlock.mode")
+	//        |...
 	Fill interface{}
 }